Geodiff integration

+++ app/inpututils.cpp
@@ -260,7 +260,7 @@ QString InputUtils::filesToString( QList<MerginFile> files )
   return resultList.join( ", " );
 }
 
-bool InputUtils::cpDir( const QString &srcPath, const QString &dstPath )
+bool InputUtils::cpDir( const QString &srcPath, const QString &dstPath, bool onlyDiffable )
 {
   bool result  = true;
   QDir parentDstDir( QFileInfo( dstPath ).path() );
@@ -279,22 +279,25 @@ bool InputUtils::cpDir( const QString &srcPath, const QString &dstPath )
     {
       if ( !cpDir( srcItemPath, dstItemPath ) )
       {
-        log( QString( "Cannot copy a dir from %1 to %2" ).arg( srcItemPath ).arg( dstItemPath ) );
+        log( "cpDir", QString( "Cannot copy a dir from %1 to %2" ).arg( srcItemPath ).arg( dstItemPath ) );
         result = false;
       }
     }
     else if ( info.isFile() )
     {
+      if ( onlyDiffable && !MerginApi::isFileDiffable( info.fileName() ) )
+        continue;
+
       if ( !QFile::copy( srcItemPath, dstItemPath ) )
       {
         if ( !QFile::remove( dstItemPath ) )
         {
-          log( QString( "Cannot remove a file from %1" ).arg( dstItemPath ) );
+          log( "cpDir", QString( "Cannot remove a file from %1" ).arg( dstItemPath ) );
           result =  false;
         }
         else if ( !QFile::copy( srcItemPath, dstItemPath ) )
         {
-          log( QString( "Cannot overwrite a file %1 with %2" ).arg( dstItemPath ).arg( dstItemPath ) );
+          log( "cpDir", QString( "Cannot overwrite a file %1 with %2" ).arg( dstItemPath ).arg( dstItemPath ) );
           result =  false;
         }
 
@@ -303,7 +306,7 @@ bool InputUtils::cpDir( const QString &srcPath, const QString &dstPath )
     }
     else
     {
-      log( QString( "Unhandled item %1 in cpDir" ).arg( info.filePath() ) );
+      log( "cpDir", QString( "Unhandled item %1 in cpDir" ).arg( info.filePath() ) );
     }
   }
   return result;
@@ -324,11 +327,11 @@ QString InputUtils::renameWithDateTime( const QString &srcPath, const QDateTime 
   return QString();
 }
 
-void InputUtils::log( const QString &msg, const QString &info )
+void InputUtils::log( const QString &topic, const QString &info )
 {
   QString logFilePath;
   QByteArray data;
-  data.append( QString( "%1 - %2. %3\n" ).arg( QDateTime().currentDateTimeUtc().toString( Qt::ISODateWithMs ) ).arg( msg ).arg( info ) );
+  data.append( QString( "%1 %2: %3\n" ).arg( QDateTime().currentDateTimeUtc().toString( Qt::ISODateWithMs ) ).arg( topic ).arg( info ) );
 
   qDebug() << data;
   appendLog( data, sLogFile );
